+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 
 - `rcli` alias for `reduct-cli`, [PR-10](https://github.com/reduct-storage/reduct-cli/pull/10)
-- Options for server URL and API token to alias add command, [PR-11](https://github.com/reduct-storage/reduct-cli/pull/11)
+- Options for server URL and API token to alias add
+  command, [PR-11](https://github.com/reduct-storage/reduct-cli/pull/11)
+- `rcl bucket ls` to list buckets, [PR-13](https://github.com/reduct-storage/reduct-cli/pull/13)
 
 ### Changed: