+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Options for server URL and API token to alias add
   command, [PR-11](https://github.com/reduct-storage/reduct-cli/pull/11)
 - `rcl bucket ls` to list buckets, [PR-13](https://github.com/reduct-storage/reduct-cli/pull/13)
+- `rcl bucket show` to browse a bucket, [PR-14](https://github.com/reduct-storage/reduct-cli/pull/14)
 
 ### Changed: